requirements.txt
Flask
pymupdf
python-docx
